chore: refactored handle-check-proof-validity into sub-functions
